requirements.txt
langchain
langchain-community
langchain-ollama
pypdf
python-docx
openpyxl
unstructured[docx,excel]
tiktoken
chromadb
python-dotenv
sentence-transformers
docx2txt